# metrics: Lazily instantiate Histogram Sample