--- a/metrics/registry.go
+++ b/metrics/registry.go
@@ -331,7 +331,10 @@ func (r *rootRegistry) Timer(name string, tags ...Tag) metrics.Timer {
 }
 
 func (r *rootRegistry) Histogram(name string, tags ...Tag) metrics.Histogram {
-	return r.HistogramWithSample(name, DefaultSample(), tags...)
+	// Use lazy instantiation to avoid allocating a Sample when getting an existing Histogram.
+	return r.registry.GetOrRegister(r.registerMetric(name, tags), func() metrics.Histogram {
+		return metrics.NewHistogram(DefaultSample())
+	}).(metrics.Histogram)
 }
 
 func (r *rootRegistry) HistogramWithSample(name string, sample metrics.Sample, tags ...Tag) metrics.Histogram {