Currently, calling registry.Histogram(...) allocates a DefaultSample() on every call. The underlying gometrics library supports lazy instantiation on the histogram itself (which is used in GetOrRegisterHistogram) but not the Sample.

This PR updates our implementation to use GetOrRegister directly, deferring sample creation to when the registry decides to allocate a new metric object.

Benchmark comparison:
```
name                                              old time/op    new time/op    delta
Histogram/Histogram_with_cached_Tag-16              3.25µs ± 1%    0.73µs ± 2%  -77.42%  (p=0.000 n=11+10)
Histogram/Histogram_with_NewTag-16                  3.55µs ± 1%    1.01µs ± 2%  -71.62%  (p=0.000 n=12+10)

name                                              old alloc/op   new alloc/op   delta
Histogram/Histogram_with_cached_Tag-16              18.8kB ± 0%     0.2kB ± 0%  -99.02%  (p=0.000 n=12+11)
Histogram/Histogram_with_NewTag-16                  18.8kB ± 0%     0.2kB ± 0%  -98.91%  (p=0.000 n=12+11)

name                                              old allocs/op  new allocs/op  delta
Histogram/Histogram_with_cached_Tag-16                8.00 ± 0%      4.00 ± 0%  -50.00%  (p=0.000 n=12+11)
Histogram/Histogram_with_NewTag-16                    11.0 ± 0%       7.0 ± 0%  -36.36%  (p=0.000 n=12+11)
```

Benchmark output on this branch:
```
BenchmarkHistogram
BenchmarkHistogram/cached_Histogram
BenchmarkHistogram/cached_Histogram-16                            	 3912604	       299.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogram/HistogramWithSample_with_cached_Tag
BenchmarkHistogram/HistogramWithSample_with_cached_Tag-16         	 1531075	       784.1 ns/op	     208 B/op	       5 allocs/op
BenchmarkHistogram/Histogram_with_cached_Tag
BenchmarkHistogram/Histogram_with_cached_Tag-16                   	 1646476	       734.8 ns/op	     184 B/op	       4 allocs/op
BenchmarkHistogram/HistogramWithSample_with_NewTag
BenchmarkHistogram/HistogramWithSample_with_NewTag-16             	 1000000	      1041 ns/op	     229 B/op	       8 allocs/op
BenchmarkHistogram/Histogram_with_NewTag
BenchmarkHistogram/Histogram_with_NewTag-16                       	 1000000	      1003 ns/op	     205 B/op	       7 allocs/op
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/palantir/pkg/227)
<!-- Reviewable:end -->